+++ pybamm/parameters/parameter_values.py
@@ -108,12 +108,14 @@ class ParameterValues:
             pybamm.citations.register(citation)
 
     @staticmethod
-    def create_from_bpx(filename):
+    def create_from_bpx(filename, target_soc=1):
         """
         Parameters
         ----------
         filename: str
             The filename of the bpx file
+        target_soc : float, optional
+            Target state of charge. Must be between 0 and 1. Default is 1.
 
         Returns
         -------
@@ -121,10 +123,21 @@ class ParameterValues:
             A parameter values object with the parameters in the bpx file
 
         """
-        from bpx import parse_bpx_file
+        if target_soc < 0 or target_soc > 1:
+            raise ValueError("Target SOC should be between 0 and 1")
+
+        from bpx import parse_bpx_file, get_electrode_concentrations
         from .bpx import bpx_to_param_dict
+
+        # parse bpx
         bpx = parse_bpx_file(filename)
         pybamm_dict = bpx_to_param_dict(bpx)
+
+        # get initial concentrations based on SOC
+        c_n_init, c_p_init = get_electrode_concentrations(target_soc, bpx)
+        pybamm_dict["Initial concentration in negative electrode [mol.m-3]"] = c_n_init
+        pybamm_dict["Initial concentration in positive electrode [mol.m-3]"] = c_p_init
+
         return pybamm.ParameterValues(pybamm_dict)
 
     def __getitem__(self, key):
